updates entire structure
