+++ frontend/src/components/About.js
@@ -1,0 +1,17 @@
+const About = () => {
+  return (
+    <div>
+      <div>
+        This is the About page
+      </div>
+      <div>
+        this page will 1. describe what TARO is, 2. explain the proposal system 3. walk users through the process of getting metamask, connecting to metamask and getting validated
+
+        the link to install metamask
+        https://metamask.io/download.html
+      </div>
+    </div>
+  );
+};
+
+export default About;
