+++ frontend/src/components/CreateProposal.js
@@ -1,0 +1,9 @@
+const CreateProposal = () => {
+  return (
+    <div>
+      Create a proposal
+    </div>
+  );
+};
+
+export default CreateProposal;
