--- frontend/src/components/FarmTaro.js
@@ -1,9 +1,0 @@
-const FarmTaro = () => {
-  return (
-    <div>
-      FarmTaro
-    </div>
-  );
-};
-
-export default FarmTaro;
