--- frontend/src/components/Footer.js
@@ -1,9 +1,0 @@
-const Footer = () => {
-  return (
-    <div>
-      When this component renders it makes the whole app stink because it is the foot-er.
-    </div>
-  );
-};
-
-export default Footer
