--- frontend/src/components/Governance.js
@@ -1,9 +1,0 @@
-const Governance = () => {
-  return (
-    <div>
-      Cinderella can't get access to her ETH because on the way home from the metamasquerade ball she lost the slipper in which she had been hiding her mnemonic.
-    </div>
-  );
-};
-
-export default Governance;
