--- frontend/src/components/InstallMetamask.js
@@ -1,11 +1,0 @@
-const InstallMetamask = () => {
-
-    return (
-      <div>
-        this page will have: 1. video about metamask, 2. quiz about metamask, 3. a link to install MetaMask
-        https://metamask.io/download.html
-      </div>
-    );
-};
-
-export default InstallMetamask;
