--- frontend/src/components/Proposal.js
@@ -1,9 +1,0 @@
-const Proposal = () => {
-  return (
-    <div>
-      Vitalik Buterin gave the keynote address at a conference lately and said, "I know that you all hate the high network expenses right now, but you have to do your part.  For instance, you can cut down on gas just by avoiding beans."
-    </div>
-  );
-};
-
-export default Proposal;
