+++ frontend/src/components/Quiz.js
@@ -2,6 +2,9 @@ const Quiz = () => {
   return (
     <div>
       <div>
+        This will be the quiz about Queretaro that allows people to get validated to make and vote on proposals
+      </div>
+      <div>
         Knock Knock
       </div>
       <div>
