--- frontend/src/components/Quizes.js
@@ -1,7 +1,0 @@
-const Quizes = () => {
-  return (
-    <div>
-      Quizes
-    </div>
-  );
-};
