--- frontend/src/components/Rewards.js
@@ -1,9 +1,0 @@
-const Rewards = () => {
-  return (
-    <div>
-      Rewards
-    </div>
-  );
-};
-
-export default Rewards;
