--- frontend/src/components/Signup.js
@@ -1,14 +1,0 @@
-const Signup = () => {
-  return (
-    <div>
-      <div>
-        You: Ha ha!  Look at you, component.  You are so boring and lazy.  You have no style and you do nothing.
-      </div>
-      <div>
-        Component:  Let's see who is laughing in a week when I am broken in a hundred different ways.  You will be begging for boring and lazy.
-      </div>
-    </div>
-  );
-};
-
-export default Signup;
