--- frontend/src/components/Voting.js
@@ -1,9 +1,0 @@
-const Voting = () => {
-  return (
-    <div>
-      Voting
-    </div>
-  );
-};
-
-export default Voting;
